Organize API responses
We have too many of these classes for us to not organize them.

This commit does a basic pass to split them up into a few basic categories. It improves things substantially.
