+++ Refresh.GameServer/Documentation/DocumentationService.cs
@@ -3,6 +3,7 @@ using System.Reflection;
 using Bunkum.Core.Services;
 using NotEnoughLogs;
 using Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response;
+using Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response.Data;
 
 namespace Refresh.GameServer.Documentation;
