+++ Refresh.GameServer/Endpoints/ApiV3/ApiTypes/ApiEmptyResponse.cs
@@ -1,4 +1,6 @@
-namespace Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response;
+using Refresh.GameServer.Endpoints.ApiV3.DataTypes;
+
+namespace Refresh.GameServer.Endpoints.ApiV3.ApiTypes;
 
 [JsonObject(NamingStrategyType = typeof(CamelCaseNamingStrategy))]
 public class ApiEmptyResponse : IApiResponse
