+++ Refresh.GameServer/Endpoints/ApiV3/DataTypes/Response/ApiContestResponse.cs
@@ -1,4 +1,6 @@
 using Refresh.GameServer.Authentication;
+using Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response.Levels;
+using Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response.Users;
 using Refresh.GameServer.Types.Contests;
 using Refresh.GameServer.Types.Data;
