+++ Refresh.GameServer/Endpoints/ApiV3/DataTypes/Response/Levels/ApiGameSkillRewardResponse.cs
@@ -1,7 +1,7 @@
 ﻿using Refresh.GameServer.Types.Data;
 using Refresh.GameServer.Types.Levels.SkillRewards;
 
-namespace Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response;
+namespace Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response.Levels;
 
 [JsonObject(NamingStrategyType = typeof(CamelCaseNamingStrategy))]
 public class ApiGameSkillRewardResponse : IApiResponse, IDataConvertableFrom<ApiGameSkillRewardResponse, GameSkillReward>
