+++ Refresh.GameServer/Endpoints/ApiV3/DataTypes/Response/Users/ApiAuthenticationResponse.cs
@@ -1,7 +1,7 @@
 #nullable disable
 using JetBrains.Annotations;
 
-namespace Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response;
+namespace Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response.Users;
 
 [JsonObject(NamingStrategyType = typeof(CamelCaseNamingStrategy), ItemNullValueHandling = NullValueHandling.Ignore)]
 public class ApiAuthenticationResponse : IApiAuthenticationResponse
