+++ Refresh.GameServer/Endpoints/ApiV3/DataTypes/Response/Users/ApiResetPasswordResponse.cs
@@ -1,5 +1,5 @@
 #nullable disable
-namespace Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response;
+namespace Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response.Users;
 
 [JsonObject(NamingStrategyType = typeof(CamelCaseNamingStrategy))]
 public class ApiResetPasswordResponse : IApiAuthenticationResponse
