+++ Refresh.GameServer/Endpoints/ApiV3/DataTypes/Response/Users/Rooms/ApiGameRoomPlayerResponse.cs
@@ -1,7 +1,7 @@
 using Refresh.GameServer.Types.Data;
 using Refresh.GameServer.Types.Matching;
 
-namespace Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response;
+namespace Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response.Users.Rooms;
 
 [JsonObject(NamingStrategyType = typeof(CamelCaseNamingStrategy))]
 public class ApiGameRoomPlayerResponse : IApiResponse, IDataConvertableFrom<ApiGameRoomPlayerResponse, GameRoomPlayer>
