+++ Refresh.GameServer/Endpoints/ApiV3/ResourceApiEndpoints.cs
@@ -12,6 +12,7 @@ using Refresh.GameServer.Database;
 using Refresh.GameServer.Endpoints.ApiV3.ApiTypes;
 using Refresh.GameServer.Endpoints.ApiV3.ApiTypes.Errors;
 using Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response;
+using Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response.Data;
 using Refresh.GameServer.Importing;
 using Refresh.GameServer.Types.Assets;
 using Refresh.GameServer.Types.Data;
