+++ Refresh.GameServer/Endpoints/ApiV3/ReviewApiEndpoints.cs
@@ -7,6 +7,7 @@ using Refresh.GameServer.Documentation.Attributes;
 using Refresh.GameServer.Endpoints.ApiV3.ApiTypes;
 using Refresh.GameServer.Endpoints.ApiV3.ApiTypes.Errors;
 using Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response;
+using Refresh.GameServer.Endpoints.ApiV3.DataTypes.Response.Levels;
 using Refresh.GameServer.Extensions;
 using Refresh.GameServer.Types.Data;
 using Refresh.GameServer.Types.Levels;
